requirements.txt
flask
google-auth
google-auth-oauthlib
google-api-python-client
requests
websocket-client>=1.6.0
werkzeug
python-dotenv
# 假設 Gemini API 有官方 Python SDK
# gemini-flash-lite-sdk
